Fix to #10813 - Query: correlated collection optimization is broken for queries containing top level First/FirstOrDefault/Single/SingleOrDefault
Problem was that correlated collection optimization rewrite wasn't properly handling result operators that could be part of the parent query model. Parent QM is cloned, rewritten and then used as part of the inner collection query.

Fix is to properly compensate for result operators that change the cardinality of the result (specifically First/Single/Last), similarly to what we do in the include pipeline.