Fix to #10828 - Query: include doesn't work for queries with order by and group by
Problem was that when rewriting collection query for case with group result operator, we add order by key element, but we were not adding it as first orderby term. This would cause outer elements arrive in different order than inner, so include would not associate them together. Fix is to apply order by always as first element.